refactor: 대형 파일 1차 분할 (Android 5 + iOS 1)
Android:
- AlarmSettingsCard(1882) → +AlarmRandomPromptSettings.kt +AlarmFortuneSettings.kt
- BillingPanels(1242) → +BillingCharacterPanel.kt
- SettingsScreen(804) → +SettingsScreenComponents.kt
- AlarmEditorScreen(1796) → +AlarmEditorScreenComponents.kt (메인 추가 분할 예정)
- VoiceAlarmApp(958) → +VoiceAlarmAppHelpers.kt
- SnoozeIntervals private→internal (분할 경계 넘는 참조 해소)
iOS:
- BillingPanel(816) → +BillingPanelComponents.swift

top-level private→internal(Kotlin)/internal(Swift) 가시성만 조정. 동작/디자인 불변.
(import 정리는 후속 커밋)
