+++ apps/android-native/app/src/main/java/com/alarmtalk/app/ui/app/VoiceAlarmApp.kt
@@ -788,171 +788,3 @@ internal fun VoiceAlarmApp(viewModel: MainViewModel = viewModel()) {
     }
 }
 
-private fun Context.openWebUrl(url: String) {
-    runCatching {
-        startActivity(
-            Intent(Intent.ACTION_VIEW, Uri.parse(url)).addFlags(Intent.FLAG_ACTIVITY_NEW_TASK),
-        )
-    }
-}
-
-private fun buildGoogleSignInOptions(requestIdToken: Boolean = false): GoogleSignInOptions {
-    val builder = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-        .requestEmail()
-    if (requestIdToken) {
-        val clientId = BuildConfig.VOICE_ALARM_GOOGLE_WEB_CLIENT_ID
-        if (clientId.isNotBlank()) {
-            builder.requestIdToken(clientId)
-        }
-    }
-    return builder.build()
-}
-
-private object AppRoute {
-    const val Settings = "settings"
-    const val MemberManagement = "members"
-    const val FamilyTargetModeArg = "familyTargetMode"
-    const val AlarmCreate = "alarm/create/{$FamilyTargetModeArg}"
-    const val AlarmIdArg = "alarmId"
-    const val AlarmEdit = "alarm/edit/{$AlarmIdArg}"
-
-    fun alarmCreate(familyTargetMode: Boolean): String = "alarm/create/$familyTargetMode"
-    fun alarmEdit(alarmId: String): String = "alarm/edit/${Uri.encode(alarmId)}"
-}
-
-private val NativeTab.route: String
-    get() = when (this) {
-        NativeTab.Home -> "home"
-        NativeTab.Voices -> "voices"
-        NativeTab.Alarms -> "alarms"
-        NativeTab.People -> "people"
-        NativeTab.Messages -> "messages"
-        NativeTab.Growth -> "growth"
-        NativeTab.Billing -> "billing"
-    }
-
-private fun String?.toNativeTab(): NativeTab? =
-    NativeTab.values().firstOrNull { it.route == this }
-
-private fun alarmPermissionRequiredMessage(target: PermissionTarget): String = when (target) {
-    PermissionTarget.Notifications -> "알람 화면과 종료 버튼을 표시하려면 알림 권한이 필요해요."
-    PermissionTarget.ExactAlarms -> "정해진 시간에 울리려면 정확한 알람 권한이 필요해요."
-    PermissionTarget.FullScreenIntent -> "잠금화면 위에 알람 화면을 띄우려면 전체 화면 알람 권한을 켜 주세요."
-    PermissionTarget.RecordAudio -> "음성을 녹음하려면 마이크 권한이 필요해요."
-}
-
-private fun NavHostController.navigateTopLevelTab(tab: NativeTab) {
-    navigate(tab.route) {
-        popUpTo(NativeTab.Home.route) {
-            saveState = true
-        }
-        launchSingleTop = true
-        restoreState = true
-    }
-}
-
-private fun NavHostController.navigateHomeClearingStack() {
-    if (currentDestination == null) return
-    navigate(NativeTab.Home.route) {
-        popUpTo(NativeTab.Home.route)
-        launchSingleTop = true
-    }
-}
-
-private fun NavHostController.popBackStackOrHome() {
-    if (!popBackStack()) {
-        navigateHomeClearingStack()
-    }
-}
-
-private suspend fun signOutGoogleAccount(context: Context) {
-    GoogleSignIn
-        .getClient(context.applicationContext, buildGoogleSignInOptions())
-        .signOut()
-        .awaitCompletion()
-}
-
-private suspend fun revokeGoogleAccountAccess(context: Context) {
-    GoogleSignIn
-        .getClient(context.applicationContext, buildGoogleSignInOptions())
-        .revokeAccess()
-        .awaitCompletion()
-}
-
-private suspend fun Task<Void>.awaitCompletion() {
-    suspendCancellableCoroutine { continuation ->
-        addOnCompleteListener { task ->
-            if (!continuation.isActive) return@addOnCompleteListener
-            when {
-                task.isCanceled -> continuation.resumeWithException(
-                    CancellationException("Google task was cancelled"),
-                )
-                task.isSuccessful -> continuation.resume(Unit)
-                else -> continuation.resumeWithException(
-                    task.exception ?: IllegalStateException("Google task failed"),
-                )
-            }
-        }
-    }
-}
-
-private sealed interface AuthRoute {
-    data object Landing : AuthRoute
-    data class Auth(val mode: AuthMode) : AuthRoute
-}
-
-private enum class MessageSeverity { Success, Error, Info }
-
-private data class PlanGateDialogState(
-    val message: String,
-    val confirmLabel: String = "이용권 보기",
-)
-
-private fun messageSeverity(text: String): MessageSeverity = when {
-    "실패" in text || "못했어요" in text || "오류" in text -> MessageSeverity.Error
-    "했어요" in text || "었어요" in text || "완료" in text -> MessageSeverity.Success
-    else -> MessageSeverity.Info
-}
-
-@Composable
-private fun PrettySnackbar(message: String) {
-    val severity = messageSeverity(message)
-    val scheme = MaterialTheme.colorScheme
-    val containerColor = when (severity) {
-        MessageSeverity.Error -> scheme.error
-        MessageSeverity.Success -> scheme.tertiary
-        MessageSeverity.Info -> scheme.secondaryContainer
-    }
-    val contentColor = when (severity) {
-        MessageSeverity.Error -> scheme.onError
-        MessageSeverity.Success -> scheme.onTertiary
-        MessageSeverity.Info -> scheme.onSecondaryContainer
-    }
-    val iconVector = when (severity) {
-        MessageSeverity.Error -> Icons.Outlined.ErrorOutline
-        MessageSeverity.Success -> Icons.Outlined.CheckCircle
-        MessageSeverity.Info -> Icons.Outlined.Info
-    }
-    Snackbar(
-        modifier = Modifier
-            .padding(horizontal = 16.dp, vertical = 8.dp)
-            .shadow(elevation = 12.dp, shape = RoundedCornerShape(20.dp), clip = false),
-        shape = RoundedCornerShape(20.dp),
-        containerColor = containerColor,
-        contentColor = contentColor,
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                imageVector = iconVector,
-                contentDescription = null,
-                tint = contentColor,
-                modifier = Modifier.size(22.dp),
-            )
-            Spacer(Modifier.width(12.dp))
-            Text(
-                text = message,
-                fontWeight = FontWeight.Medium,
-            )
-        }
-    }
-}
